pyproject.toml
[project]
name = "personaltranscribe"
version = "0.1.0"
description = "Voice transcription application with editing and PDF export for legal use"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "faster-whisper>=1.0.0",
    "PyQt6>=6.5.0,<6.10.0",
    "PyQt6-Qt6>=6.5.0,<6.10.0",
    "pyqtgraph>=0.13.0",
    "pydub>=0.25.0",
    "fpdf2>=2.7.0",
    "numpy>=1.24.0",
    "soundfile>=0.12.0",
    "python-docx>=1.0.0",
    "nvidia-cudnn-cu12>=9.17.1.4",
    "openai>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
